finish

+++ src/main/java/edu/epam/course/command/impl/TeacherDeleteCommand.java
@@ -1,17 +1,15 @@
 package edu.epam.course.command.impl;
 
-import edu.epam.course.command.Command;
-import edu.epam.course.command.PagePath;
-import edu.epam.course.command.RequestAttribute;
-import edu.epam.course.command.RequestParameter;
-import edu.epam.course.command.Router;
+import edu.epam.course.command.*;
 import edu.epam.course.exception.ServiceException;
 import edu.epam.course.model.service.TeacherService;
 import edu.epam.course.util.IdUtil;
+import edu.epam.course.validator.TeacherValidator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -37,14 +35,22 @@ public class TeacherDeleteCommand implements Command {
     @Override
     public Router execute(HttpServletRequest request) {
         String[] teachersIdString = request.getParameterValues(RequestParameter.TEACHER_ID);
+        HttpSession session = request.getSession();
         List<Long> teachersId = new ArrayList<>();
         Router router = new Router();
+        boolean dataCorrect = true;
         try {
-            for (int i = 0; i < teachersIdString.length; i++) {
-                Long teacherId = IdUtil.stringToLong(teachersIdString[i]);
-                teachersId.add(teacherId);
+            if (!TeacherValidator.isValidTeachersId(teachersIdString)) {
+                session.setAttribute(SessionAttribute.ERROR_SELECT_TEACHER_ID, true);
+                dataCorrect = false;
             }
-            teacherService.deleteTeachers(teachersId);
+            if (dataCorrect) {
+                for (int i = 0; i < teachersIdString.length; i++) {
+                    Long teacherId = IdUtil.stringToLong(teachersIdString[i]);
+                    teachersId.add(teacherId);
+                }
+                teacherService.deleteTeachers(teachersId);
+            }
             router.setPagePath(PagePath.SHOW_ALL_TEACHERS.getServletPath());
         } catch (ServiceException | NumberFormatException e) {
             logger.error(e);
